+++ include/toland_flight/FlightDetector.hpp
@@ -1,12 +1,23 @@
+/// Copyright (C) 2021 Martin Scheiber, Control of Networked Systems, University of Klagenfurt, Austria.
+///
+/// All rights reserved.
+///
+/// This software is licensed under the terms of the BSD-2-Clause-License with
+/// no commercial use allowed, the full terms of which are made available
+/// in the LICENSE file. No license in patents is granted.
+///
+/// You can contact the author at <martin.scheiber@ieee.org>
+
 #ifndef FLIGHTDETECTOR_HPP
 #define FLIGHTDETECTOR_HPP
 
 #include <ros/ros.h>
 #include <sensor_msgs/Imu.h>
 #include <sensor_msgs/Range.h>
+#include <std_srvs/Trigger.h>
 
-#include "flatness/sensors.h"
-#include "flatness/mathematics.h"
+#include "utils/sensors.h"
+#include "utils/mathematics.h"
 
 namespace toland
 {
@@ -15,6 +26,8 @@ namespace toland
    */
   class FlightDetector
   {
+  private:
+    typedef utils::sensors::imuData ImuData_t ;
 
   private:
     /// ROS NODE HANDLES
@@ -24,13 +37,15 @@ namespace toland
     ros::Subscriber sub_imu_;                                     //!< IMU topic subscriber
     ros::Subscriber sub_lrf_;                                     //!< LRF topic subscriber
 
+    ros::ServiceServer srv_to_;                                   //!< Takeoff service
+
     /// Measurement buffers
-    std::vector<imuData> imu_data_buffer_;                        //!< buffer for IMU measurements
+    std::vector<ImuData_t> imu_data_buffer_;                       //!< buffer for IMU measurements
 
     /// ROS Static Parameters
     double k_sensor_readings_window_s_ {1.0};
     double k_angle_threshold_deg_ {10.0};
-
+    std::vector<double> k_R_IP = {1,0,0,0,1,0,0,0,1};
 
     /// ROS Callback Functions
 
@@ -45,7 +60,17 @@ namespace toland
     /// \brief lrfCallback
     /// \param msg
     ///
-    void lrfCallback(const sensor_msgs::Range::ConstPtr& msg);    //!<
+    void lrfCallback(const sensor_msgs::Range::ConstPtr& msg);
+
+    bool takeoffHandler(std_srvs::Trigger::Request& req,
+                        std_srvs::Trigger::Response& res);
+
+    ///
+    /// \brief checkFlatness
+    /// \return
+    /// \author Alessandro Fornasier
+    ///
+    bool checkFlatness();
 
 public:
     FlightDetector();