+++ CHANGELOG.md
@@ -5,5 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
+### Added
+- Added ROS package structure
+- Added IMU flatness detection from AMAZE Autonomy.
+- Added service for takeoff request using `std_srvs/Trigger.srv`
 
-[Unreleased]: https://gitlab.aau.at/.../-/compare/develop...main
+### Changed
+- Added namespaces for flatness detection (utils) from AMAZE autonomy.
+- Updated license headers in utils module.
+
+[Unreleased]: hhttps://gitlab.aau.at/aau-cns/ros_pkgs/toland_flight/-/compare/develop...main